+++ src/Hasklepias/Cohort.hs
@@ -5,6 +5,7 @@ Copyright   : (c) NoviSci, Inc 2020
 License     : BSD3
 Maintainer  : bsaul@novisci.com
 -}
+{-# OPTIONS_HADDOCK hide #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE TupleSections #-}
@@ -22,28 +23,35 @@ module Hasklepias.Cohort(
     , specifyCohort
     , makeObsUnitFeatures
     , evalCohort
+
+    -- ** Criteria
     , module Hasklepias.Cohort.Criteria
+
+    -- ** Index
     , module Hasklepias.Cohort.Index
 ) where
 
-import Prelude                  ( Eq, Show, Bool, init )
-import GHC.Num                  ( Num((+)), Natural )
-import Data.Aeson               ( FromJSON, ToJSON, ToJSONKey )
-import Data.Foldable ( Foldable(length) )
-import Data.Function            ( ($) )
-import Data.Functor             ( Functor(fmap) )
-import Data.Maybe               ( Maybe(..), catMaybes )
-import Data.List                ( zipWith, zip, replicate )
-import qualified Data.List.NonEmpty as NE
-import Data.Map.Strict as Map   ( toList, fromListWith )
-import Data.Text                ( Text )
-import GHC.Generics             ( Generic )
-import Hasklepias.Cohort.Index
+import GHC.Num                              ( Num((+)), Natural )
+import Data.Aeson                           ( FromJSON, ToJSON, ToJSONKey )
+import Data.Bool                            ( Bool )
+import Data.Eq                              ( Eq )
+import Data.Foldable                        ( Foldable(length) )
+import Data.Function                        ( ($) )
+import Data.Functor                         ( Functor(fmap) )
+import Data.Maybe                           ( Maybe(..), catMaybes )
+import Data.List                            ( zipWith, replicate )
+import Data.List.NonEmpty                   ( NonEmpty(..), zip, fromList, nonEmpty )
+import Data.Map.Strict as Map               ( toList, fromListWith )
+import Data.Text                            ( Text )
+import GHC.Generics                         ( Generic )
+import GHC.Show                             ( Show )
+import Hasklepias.Cohort.Index              ( makeIndex, Index(..) )
 import Hasklepias.Cohort.Criteria
-import FeatureCompose
 
-
+-- | A subject identifier. Currently, simply @Text@.
 type ID = Text
+
+-- | A subject is just a pair of @ID@ and data.
 newtype Subject d = MkSubject (ID, d)
     deriving (Eq, Show, Generic)
 
@@ -52,71 +60,91 @@ instance Functor Subject where
 
 instance (FromJSON d) => FromJSON (Subject d) where
 
-newtype Population d = MkPopulation [Subject d]
+-- | A population is a list of @'Subject'@s
+newtype Population d = MkPopulation  [Subject d] 
     deriving (Eq, Show, Generic)
-
-instance (FromJSON d) => FromJSON (Population d) where
 
 instance Functor Population where
     fmap f (MkPopulation x) = MkPopulation (fmap (fmap f) x)
 
+instance (FromJSON d) => FromJSON (Population d) where
+
+-- | An observational unit is what a subject may be transformed into.
 newtype ObsUnit d = MkObsUnit (ID, d)
     deriving (Eq, Show, Generic)
 
 instance (ToJSON d) => ToJSON (ObsUnit d) where
 
-newtype Cohort d = MkCohort (AttritionInfo, [ObsUnit d])
+-- | A cohort is a list of observational units along with @'AttritionInfo'@ 
+-- regarding the number of subjects excluded by the @'Criteria'@. 
+newtype Cohort d = MkCohort (Maybe AttritionInfo, [ObsUnit d])
     deriving (Eq, Show, Generic)
 
 instance (ToJSON d) => ToJSON (Cohort d) where
 
+-- | Unpacks a @'Population'@ to a list of subjects.
 getPopulation :: Population d -> [Subject d]
 getPopulation (MkPopulation x) = x
 
+-- | Gets the data out of  a @'Subject'@.
 getSubjectData :: Subject d -> d
 getSubjectData (MkSubject (_, x)) = x
 
+-- | Tranforms a @'Subject'@ into a @'ObsUnit'@.
 makeObsUnitFeatures :: (d1 -> d0) -> Subject d1 -> ObsUnit d0
 makeObsUnitFeatures f (MkSubject (id, dat)) = MkObsUnit (id, f dat)
 
--- makeCohort :: (d1 -> d0) -> Population d1 -> Cohort d0
--- makeCohort f (MkPopulation x) = MkCohort (fmap (makeObsUnitFeatures f) x)
-
+-- | A cohort specification consist of two functions: one that transforms a subject's
+-- input data into a @'Criteria'@ and another that transforms a subject's input data
+-- into the desired return type.
 data CohortSpec d1 d0 = MkCohortSpec
         { runCriteria:: d1 -> Criteria
         -- (Feature b (Index i a))
         , runFeatures:: d1 -> d0 }
 
+-- | Creates a @'CohortSpec'@.
 specifyCohort :: (d1 -> Criteria) -> (d1 -> d0) ->  CohortSpec d1 d0
 specifyCohort = MkCohortSpec
 
+-- | Evaluates the @'runCriteria'@ of a @'CohortSpec'@ on a @'Population'@ to 
+-- return a list of @Subject Criteria@ (one per subject in the population). 
 evalCriteria :: CohortSpec d1 d0 -> Population d1 -> [Subject Criteria]
 evalCriteria (MkCohortSpec runCrit _) (MkPopulation pop) = fmap (fmap runCrit) pop
 
+-- | Convert a list of @Subject Criteria@ into a list of @Subject CohortStatus@
 evalCohortStatus :: [Subject Criteria] -> [Subject CohortStatus]
 evalCohortStatus = fmap (fmap checkCohortStatus)
 
+-- | Runs the input function which transforms a subject into an observational unit. 
+-- If the subeject is excluded, the result is @Nothing@; otherwise it is @Just@ 
+-- an observational unit.
 evalSubjectCohort :: (d1 -> d0) -> Subject CohortStatus -> Subject d1 -> Maybe (ObsUnit d0)
 evalSubjectCohort f (MkSubject (id, status)) subjData =
     case status of
         Included     -> Just $ makeObsUnitFeatures f subjData
         ExcludedBy _ -> Nothing
 
-newtype AttritionInfo = MkAttritionInfo [(CohortStatus, Natural)]
+-- | A type which collects the counts of subjects included or excluded.
+newtype AttritionInfo = MkAttritionInfo (NonEmpty (CohortStatus, Natural))
     deriving (Eq, Show, Generic)
 
+-- | Initializes @AttritionInfo@ from a @'Criteria'@.
 initAttritionInfo :: Criteria -> AttritionInfo
 initAttritionInfo x =
-    MkAttritionInfo $ zip (initStatusInfo x) (replicate (length (getCriteria x)) 0)
+    MkAttritionInfo $ zip (initStatusInfo x) 
+        (0 :| replicate (length (getCriteria x)) 0)
 
 instance ToJSON CohortStatus where
 instance ToJSON AttritionInfo where
 
-measureAttrition :: [Subject CohortStatus] -> AttritionInfo
-measureAttrition l = MkAttritionInfo $ Map.toList $
+-- | Creates an @'AttritionInfo'@ from a list of @Subject CohortStatus@. The result
+-- is @Nothing@ if the input list is empty.
+measureAttrition :: [Subject CohortStatus] -> Maybe AttritionInfo
+measureAttrition l = fmap MkAttritionInfo $ nonEmpty $ Map.toList $
      Map.fromListWith (+) $ fmap (\x -> (getSubjectData x, 1)) l
 
-evalUnits :: CohortSpec d1 d0 -> Population d1 -> (AttritionInfo, [ObsUnit d0])
+-- | The internal function to evaluate a @'CohortSpec'@ on a @'Population'@. 
+evalUnits :: CohortSpec d1 d0 -> Population d1 -> (Maybe AttritionInfo, [ObsUnit d0])
 evalUnits spec pop =
     ( measureAttrition statuses
     , catMaybes $ zipWith (evalSubjectCohort (runFeatures spec))
@@ -125,5 +153,6 @@ evalUnits spec pop =
     where crits = evalCriteria spec pop
           statuses = evalCohortStatus crits
 
+-- | Evaluates a @'CohortSpec'@ on a @'Population'@.
 evalCohort :: CohortSpec d1 d0 -> Population d1 -> Cohort d0
 evalCohort s p = MkCohort $ evalUnits s p